EID-918: Consolidate classes in a core package
Creating a package to house the major classes that relate to business concepts.

Authors: @kgarwood, @felixmarcusmillne
